Move the "money-back guarantee" paragraph
