+++ client/signup/steps/page-picker/index.tsx
@@ -3,6 +3,7 @@ import { getDIFMTieredPriceDetails, WPCOM_DIFM_LITE } from '@automattic/calypso-
 import { RazorpayHookProvider } from '@automattic/calypso-razorpay';
 import { StripeHookProvider } from '@automattic/calypso-stripe';
 import { Button } from '@automattic/components';
+import { BraveTick } from '@automattic/components/src/icons';
 import { formatCurrency } from '@automattic/number-formatters';
 import { createRequestCartProduct } from '@automattic/shopping-cart';
 import { isMobile } from '@automattic/viewport';
@@ -374,6 +375,24 @@ const Placeholder = styled.span`
 	min-width: 32px;
 `;
 
+const RefundText = styled.div`
+	display: flex;
+	align-items: center;
+	width: 100%;
+	margin-top: 24px;
+
+	@media ( max-width: 600px ) {
+		font-size: smaller;
+	}
+`;
+
+const StyledBraveTickIcon = styled( BraveTick )`
+	margin-inline-end: 6px;
+	path {
+		fill: var( --color-accent );
+	}
+`;
+
 function OneClickPurchaseModal( {
 	onClose,
 	siteSlug,
@@ -610,14 +629,23 @@ function DIFMPagePicker( props: StepProps ) {
 			isHorizontalLayout
 			isWideLayout={ false }
 			headerButton={
-				<StyledButton
-					disabled={ isProductsLoading }
-					busy={ isCheckoutButtonBusy }
-					primary
-					onClick={ submitPickedPages }
-				>
-					{ translate( 'Go to Checkout' ) }
-				</StyledButton>
+				<>
+					<StyledButton
+						disabled={ isProductsLoading }
+						busy={ isCheckoutButtonBusy }
+						primary
+						onClick={ submitPickedPages }
+					>
+						{ translate( 'Go to Checkout' ) }
+					</StyledButton>
+
+					<RefundText>
+						<StyledBraveTickIcon />
+						{ translate( '%(days)d-day money-back guarantee', {
+							args: { days: 14 },
+						} ) }
+					</RefundText>
+				</>
 			}
 			headerContent={
 				<ShoppingCartForDIFM
